Rewrite test_booking_reference_service.py
