+++ tests/test_train_data_service.py
@@ -5,7 +5,7 @@ import json
 import cherrypy
 from cherrypy.test import helper
 
-from train_data_service.train_data_service import TrainDataService
+from train_reservation.train_data_service import TrainDataService
 
 
 class TrainDataServiceTest(helper.CPWebCase):
